Detrended Quantile Mapping

This function performs bias correction of modeled series based on observed data by the Detrended Quantile Mapping (DQM) method, as described by Cannon et al. (2015). Correction is performed to monthly or annual precipitation or temperature data in a single location. An independent probability distribution function is assigned to each month and to each projected period based on the Kolmogorov-Smirnov test. If annual frequency is specified, this is applied to the complete period. Each projected period considers a window whose length is equal to the number of years of the historical period and ends in the analyzed year. Correction of the historical period is performed by the Quantile Mapping (QM) method, as described by Cannon et al. (2015).

DQM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)

Arguments

obs

A column vector of monthly or annual observed data (temperature or precipitation). If monthly frequency is specified, the length of this vector is 12 times the number of observed years [12 x y_obs, 1]. If annual frequency is specified, the length of this vector is equal to the number of observed years [y_obs, 1].

mod

A column vector of monthly or annual modeled data (temperature or precipitation). If monthly frequency is specified, the length of this vector is 12 times the number of observed years [12 x y_mod, 1]. If annual frequency is specified, the length of this vector is equal to the number of observed years [y_mod, 1].

var

A flag that identifies if data are temperature or precipitation. This flag tells the getDist function if it has to discard distribution functions that allow negative numbers, and if the terms in the correction equations are multiplied/divided or added/subtracted. Temperature: var = 0; Precipitation: var = 1

frq

(Optional) A string specifying if the input is annual or monthly data. If not specified, it is set monthly as default. Monthly: frq = 'M'; Annual: frq = 'A'

pp_threshold

(Optional) A float indicating the threshold to consider physically null precipitation values.

pp_factor

(Optional) A float indicating the maximum value of the random values that replace physically null precipitation values.

Value

A column vector of monthly or annual modeled data (temperature or precipitation) corrected by the DQM method. If monthly frequency is specified, the length of this vector is 12 times the number of observed years [12 x y_mod, 1]. If annual frequency is specified, the length of this vector is equal to the number of observed years [y_mod, 1].

Details

Cannon, A. J., S. R. Sobie, and T. Q. Murdock, 2015: Bias correction of GCM precipitation by quantile mapping: How well do methods preserve changes in quantiles and extremes? J. Climate, 28(17), 6938-6959, https://doi.org/10.1175/JCLI-D-14-00754.1

Examples

DQM(obs,mod,var)
#> Error in var == 1: comparison (==) is possible only for atomic and list types
DQM(obs,mod,var,frq='A')
#> Error in var == 1: comparison (==) is possible only for atomic and list types
DQM(obs,mod,var,frq='M')
#> Error in var == 1: comparison (==) is possible only for atomic and list types
DQM(obs,mod,var,frq='M',pp_threshold=0.1)
#> Error in var == 1: comparison (==) is possible only for atomic and list types
DQM(obs,mod,var,frq='M',pp_factor=1/1000)
#> Error in var == 1: comparison (==) is possible only for atomic and list types
DQM(obs,mod,var,frq='M',pp_threshold=0.1,pp_factor=1/1000)
#> Error in var == 1: comparison (==) is possible only for atomic and list types